Fixing comments
